Syslog POSIX: Moved from posix/sys/ to posix/
