+++ api/posix/syslog.h
